requirements.txt
# Core dependencies
pandas>=1.5.0,<3.0.0
requests>=2.28.0,<3.0.0

# Development and testing dependencies
pytest>=7.0.0,<8.0.0
pytest-cov>=4.0.0,<5.0.0
pytest-mock>=3.10.0,<4.0.0
responses>=0.21.0,<1.0.0

# Code quality
black>=22.0.0,<24.0.0
flake8>=5.0.0,<7.0.0
isort>=5.10.0,<6.0.0

# Optional: For enhanced functionality
python-dotenv>=0.19.0,<2.0.0  # For .env file support 